refactor(frontend): use domain exception filename kebab case
