--- frontend/app/.server/domain/exceptions/ClientFriendlyStatusNotFoundException.ts
@@ -1,3 +1,0 @@
-import { ResourceNotFoundException } from '~/.server/domain/exceptions/ResourceNotFoundException';
-
-export class ClientFriendlyStatusNotFoundException extends ResourceNotFoundException {}
