--- frontend/app/.server/domain/exceptions/CountryNotFoundException.ts
@@ -1,3 +1,0 @@
-import { ResourceNotFoundException } from '~/.server/domain/exceptions/ResourceNotFoundException';
-
-export class CountryNotFoundException extends ResourceNotFoundException {}
