--- frontend/app/.server/domain/exceptions/FederalGovernmentInsurancePlanNotFoundException.ts
@@ -1,3 +1,0 @@
-import { ResourceNotFoundException } from '~/.server/domain/exceptions/ResourceNotFoundException';
-
-export class FederalGovernmentInsurancePlanNotFoundException extends ResourceNotFoundException {}
