--- frontend/app/.server/domain/exceptions/MaritalStatusNotFoundException.ts
@@ -1,3 +1,0 @@
-import { ResourceNotFoundException } from '~/.server/domain/exceptions/ResourceNotFoundException';
-
-export class MaritalStatusNotFoundException extends ResourceNotFoundException {}
