--- frontend/app/.server/domain/exceptions/PreferredCommunicationMethodNotFoundException.ts
@@ -1,3 +1,0 @@
-import { ResourceNotFoundException } from '~/.server/domain/exceptions/ResourceNotFoundException';
-
-export class PreferredCommunicationMethodNotFoundException extends ResourceNotFoundException {}
