--- frontend/app/.server/domain/exceptions/PreferredLanguageNotFoundException.ts
@@ -1,3 +1,0 @@
-import { ResourceNotFoundException } from '~/.server/domain/exceptions/ResourceNotFoundException';
-
-export class PreferredLanguageNotFoundException extends ResourceNotFoundException {}
