--- frontend/app/.server/domain/exceptions/ProvinceTerritoryStateNotFoundException.ts
@@ -1,3 +1,0 @@
-import { ResourceNotFoundException } from '~/.server/domain/exceptions/ResourceNotFoundException';
-
-export class ProvinceTerritoryStateNotFoundException extends ResourceNotFoundException {}
