--- frontend/app/.server/domain/exceptions/ResourceNotFoundException.ts
@@ -1,3 +1,0 @@
-import { BaseDomainException } from '~/.server/domain/exceptions/BaseDomainException';
-
-export class ResourceNotFoundException extends BaseDomainException {}
