+++ frontend/app/.server/domain/exceptions/base-domain.exception.ts
