+++ frontend/app/.server/domain/exceptions/client-friendly-status-not-found.exception.ts
@@ -1,0 +1,3 @@
+import { ResourceNotFoundException } from '~/.server/domain/exceptions/resource-not-found.exception';
+
+export class ClientFriendlyStatusNotFoundException extends ResourceNotFoundException {}
