+++ frontend/app/.server/domain/exceptions/federal-government-insurance-plan-not-found.exception.ts
@@ -1,0 +1,3 @@
+import { ResourceNotFoundException } from '~/.server/domain/exceptions/resource-not-found.exception';
+
+export class FederalGovernmentInsurancePlanNotFoundException extends ResourceNotFoundException {}
