+++ frontend/app/.server/domain/exceptions/index.ts
@@ -1,0 +1,10 @@
+export * from './base-domain.exception';
+export * from './client-friendly-status-not-found.exception';
+export * from './country-not-found.exception';
+export * from './federal-government-insurance-plan-not-found.exception';
+export * from './marital-status-not-found.exception';
+export * from './preferred-communication-method-not-found.exception';
+export * from './preferred-language-not-found.exception';
+export * from './province-territory-state-not-found.exception';
+export * from './provincial-government-insurance-plan-not-found.exception';
+export * from './resource-not-found.exception';
