+++ frontend/app/.server/domain/exceptions/preferred-language-not-found.exception.ts
@@ -1,0 +1,3 @@
+import { ResourceNotFoundException } from '~/.server/domain/exceptions/resource-not-found.exception';
+
+export class PreferredLanguageNotFoundException extends ResourceNotFoundException {}
