+++ frontend/app/.server/domain/exceptions/provincial-government-insurance-plan-not-found.exception.ts
@@ -1,3 +1,3 @@
-import { ResourceNotFoundException } from '~/.server/domain/exceptions/ResourceNotFoundException';
+import { ResourceNotFoundException } from '~/.server/domain/exceptions/resource-not-found.exception';
 
 export class ProvincialGovernmentInsurancePlanNotFoundException extends ResourceNotFoundException {}
