+++ frontend/app/.server/domain/exceptions/resource-not-found.exception.ts
@@ -1,0 +1,3 @@
+import { BaseDomainException } from '~/.server/domain/exceptions/base-domain.exception';
+
+export class ResourceNotFoundException extends BaseDomainException {}
