+++ frontend/app/.server/domain/services/preferred-language.service.ts
@@ -1,10 +1,10 @@
 import { inject, injectable } from 'inversify';
 import moize from 'moize';
 
-import { PreferredLanguageNotFoundException } from '../exceptions/PreferredLanguageNotFoundException';
 import type { ServerConfig } from '~/.server/configs';
 import { SERVICE_IDENTIFIER } from '~/.server/constants';
 import type { PreferredLanguageDto } from '~/.server/domain/dtos';
+import { PreferredLanguageNotFoundException } from '~/.server/domain/exceptions';
 import type { PreferredLanguageDtoMapper } from '~/.server/domain/mappers';
 import type { PreferredLanguageRepository } from '~/.server/domain/repositories';
 import type { LogFactory, Logger } from '~/.server/factories';
