+++ frontend/app/.server/domain/services/province-territory-state.service.ts
@@ -1,10 +1,10 @@
 import { inject, injectable } from 'inversify';
 import moize from 'moize';
 
-import { ProvinceTerritoryStateNotFoundException } from '../exceptions/ProvinceTerritoryStateNotFoundException';
 import type { ServerConfig } from '~/.server/configs';
 import { SERVICE_IDENTIFIER } from '~/.server/constants';
 import type { ProvinceTerritoryStateDto } from '~/.server/domain/dtos';
+import { ProvinceTerritoryStateNotFoundException } from '~/.server/domain/exceptions';
 import type { ProvinceTerritoryStateDtoMapper } from '~/.server/domain/mappers';
 import type { ProvinceTerritoryStateRepository } from '~/.server/domain/repositories';
 import type { LogFactory, Logger } from '~/.server/factories';
